Move route binding (used in Glob) to a trait Added tests for Route Added missing @covers
